chore: adding getAccounts query fnc by clientId
